requirements.txt
apache-airflow==2.7.0
dvc==3.30.0
pandas==2.0.0
numpy==1.24.0
pytest==7.4.0
great-expectations==0.17.0
scikit-learn==1.3.0
nltk==3.8.1
